+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 
 ### Features
 
-* add @runmq/nestjs module with decorator-based processors and publisher ([d973486](https://github.com/runmq/nestjs/commit/d973486b711545d5753f6878b73a997102522d5e))
+* add nestjs-runmq module with decorator-based processors and publisher ([d973486](https://github.com/runmq/nestjs/commit/d973486b711545d5753f6878b73a997102522d5e))